Modify setting of 'seed' for tlnise
Setting the seed for 'tlnise' has never worked. 'initTLNise' was introduced
to fix this problem but it turns out that doesn't work either. Basically, the
problem is that the 'ran2' Fortran function needs to be initialized once,
i.e. the first time it is called. But setting a seed > 0 skips this
initialization and therefore you end up with random results. For now the
solution is to pass a seed < 0 from R, which tells 'ran2' to initialize the
first time. The sign of the seed is then flipped and then everything goes
on as before (as far as I can tell).